# swagger.yml
basePath: /truphone
consumes:
- application/json
definitions:
  AgeFamily:
    description: A family with its age.
    properties:
      age:
        description: Age for the age family.
        format: int64
        type: integer
        x-go-name: Age
      country_code:
        description: the country code for this family
        maxLength: 2
        minLength: 2
        type: string
        x-go-name: CountryCode
      id:
        description: the id for this family
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      name:
        description: the name for this family
        maxLength: 30
        minLength: 3
        type: string
        x-go-name: Name
    required:
    - id
    - name
    - country_code
    - age
    title: AgeFamily
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/pkg/family
  CreateFamilyRequest:
    description: Request by CreateFamily Service.
    properties:
      family:
        $ref: '#/definitions/Family'
    required:
    - family
    title: CreateFamilyRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  CreateFamilyResponse:
    description: Response by CreateFamily Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the message for the request
        type: string
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: CreateFamilyResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  CreateMemberRequest:
    description: Request by CreateMember Service.
    properties:
      member:
        $ref: '#/definitions/Member'
    required:
    - member
    title: CreateMemberRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  CreateMemberResponse:
    description: Response by CreateMember Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the message for the request
        type: string
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: CreateMemberResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  DeleteFamilyRequest:
    description: Request by DeleteFamily Service.
    properties:
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
    required:
    - id
    title: DeleteFamilyRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  DeleteFamilyResponse:
    description: Response by DeleteFamily Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
    required:
    - id
    - code
    - message
    title: DeleteFamilyResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  DeleteMemberRequest:
    description: Request by DeleteMember Service.
    properties:
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
    required:
    - id
    title: DeleteMemberRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  DeleteMemberResponse:
    description: Response by DeleteMember Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
    required:
    - id
    - code
    - message
    title: DeleteMemberResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  Family:
    description: A family can have a name and a country.
    properties:
      country_code:
        description: the country code for this family
        maxLength: 2
        minLength: 2
        type: string
        x-go-name: CountryCode
      id:
        description: the id for this family
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      name:
        description: the name for this family
        maxLength: 30
        minLength: 3
        type: string
        x-go-name: Name
    required:
    - id
    - name
    - country_code
    title: Family
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/pkg/family
  GetFamilyRequest:
    description: Request by GetFamily Service.
    properties:
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
    required:
    - id
    title: GetFamilyRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  GetFamilyResponse:
    description: Response by GetFamily Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        $ref: '#/definitions/Family'
    required:
    - id
    - code
    - message
    - result
    title: GetFamilyResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  GetMemberRequest:
    description: Request by GetMember Service.
    properties:
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
    required:
    - id
    title: GetMemberRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  GetMemberResponse:
    description: Response by GetMember Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        $ref: '#/definitions/Member'
    required:
    - id
    - code
    - message
    - result
    title: GetMemberResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListAccumulatedFamiliesRequest:
    description: Request by ListAccumulatedFamilies Service.
    properties:
      limit:
        description: the limit for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Limit
      offset:
        description: the offset for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Offset
    title: ListAccumulatedFamiliesRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListAccumulatedFamiliesResponse:
    description: Response by ListAccumulatedFamilies Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the members for the request
        items:
          $ref: '#/definitions/AgeFamily'
        type: array
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: ListAccumulatedFamiliesResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListFamiliesRequest:
    description: Request by ListFamilies Service.
    properties:
      country_code:
        description: the country code for the request
        example: PT
        type: string
        x-go-name: CountryCode
      limit:
        description: the limit for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Limit
      offset:
        description: the offset for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Offset
    title: ListFamiliesRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListFamiliesResponse:
    description: Response by ListFamilies Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the families for the request
        items:
          $ref: '#/definitions/Family'
        type: array
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: ListFamiliesResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListFastestGrowingFamiliesRequest:
    description: Request by ListFastestGrowingFamilies Service.
    properties:
      limit:
        description: the limit for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Limit
      offset:
        description: the offset for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Offset
    title: ListFastestGrowingFamiliesRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListFastestGrowingFamiliesResponse:
    description: Response by ListFastestGrowingFamilies Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the members for the request
        items:
          $ref: '#/definitions/AgeFamily'
        type: array
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: ListFastestGrowingFamiliesResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListMembersRequest:
    description: Request by ListMembers Service.
    properties:
      family_id:
        description: the family id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: FamilyId
      limit:
        description: the limit for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Limit
      offset:
        description: the offset for the request
        example: 1
        format: uint32
        type: integer
        x-go-name: Offset
      parent_id:
        description: the parent id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: ParentId
    title: ListMembersRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  ListMembersResponse:
    description: Response by ListMembers Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        description: the members for the request
        items:
          $ref: '#/definitions/Member'
        type: array
        x-go-name: Result
    required:
    - id
    - code
    - message
    - result
    title: ListMembersResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  Member:
    description: A member can have a father, mother and spouse.
    properties:
      birthday:
        description: the birthday for this member
        example: "2012-02-03T00:04:05Z"
        format: date-time
        type: string
        x-go-name: Birthday
      family_id:
        description: the family id for this member
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: FamilyId
      father_id:
        description: the father id for this member
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: FatherId
      first_name:
        description: the first name for this member
        maxLength: 30
        minLength: 3
        type: string
        x-go-name: FirstName
      id:
        description: the id for this member
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      last_name:
        description: the last name for this member
        maxLength: 30
        minLength: 3
        type: string
        x-go-name: LastName
      middle_name:
        description: the middle name for this member
        maxLength: 30
        minLength: 3
        type: string
        x-go-name: MiddleName
      mother_id:
        description: the mother id for this member
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: MotherId
      spouse_id:
        description: the spouse id for this member
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: SpouseId
    required:
    - id
    - family_id
    - first_name
    - middle_name
    - last_name
    - birthday
    title: Member
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/pkg/family
  UpdateFamilyRequest:
    description: Request by UpdateFamily Service.
    properties:
      family:
        $ref: '#/definitions/Family'
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
    required:
    - id
    - family
    title: UpdateFamilyRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  UpdateFamilyResponse:
    description: Response by UpdateFamily Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        $ref: '#/definitions/Family'
    required:
    - id
    - code
    - message
    - result
    title: UpdateFamilyResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  UpdateMemberRequest:
    description: Request by UpdateMember Service.
    properties:
      id:
        description: the family for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      member:
        $ref: '#/definitions/Member'
    required:
    - id
    - member
    title: UpdateMemberRequest
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
  UpdateMemberResponse:
    description: Response by UpdateMember Service.
    properties:
      code:
        description: the response code for the request
        example: 200
        format: int64
        type: integer
        x-go-name: Code
      id:
        description: the id for the request
        example: 8957bf28-aea0-47de-abe0-d4c5ea593ec6
        type: string
        x-go-name: Id
      message:
        description: the message for the request
        example: success
        type: string
        x-go-name: Message
      result:
        $ref: '#/definitions/Member'
    required:
    - id
    - code
    - message
    - result
    title: UpdateMemberResponse
    type: object
    x-go-package: github.com/ricardomgoncalves/truphone_ta_go/internal/service
host: 127.0.0.1:8080
info:
  contact:
    email: ricardo.341928@hotmail.com
    name: Ricardo Goncalves
    url: http://localhost
  description: |-
    the purpose of this application is to provide a management support
    of families.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Family API.
  version: 0.0.1
paths:
  /accumulate:
    get:
      operationId: listAccumulatedFamilies
      parameters:
      - description: offset number of results to return
        format: int32
        in: query
        name: offset
        type: integer
      - description: maximum number of results to return
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/ListAccumulatedFamiliesResponse'
      summary: Lists the most accumulated aged Families.
  /families:
    get:
      operationId: listFamilies
      parameters:
      - description: offset number of results to return
        format: int32
        in: query
        name: offset
        type: integer
      - description: maximum number of results to return
        format: int32
        in: query
        name: limit
        type: integer
      - description: country code to filter
        format: string
        in: query
        name: country
        type: string
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/ListFamiliesResponse'
      summary: Lists Families.
    post:
      operationId: createFamily
      parameters:
      - description: family request
        in: body
        name: family
        required: true
        schema:
          $ref: '#/definitions/CreateFamilyRequest'
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/CreateFamilyResponse'
      summary: Creates a new Family.
  /families/{id}:
    delete:
      operationId: deleteFamily
      parameters:
      - in: path
        name: id
        required: true
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/DeleteFamilyResponse'
      summary: Deletes a Family.
    get:
      operationId: getFamily
      parameters:
      - in: path
        name: id
        required: true
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/GetFamilyResponse'
      summary: Gets a Family.
    put:
      operationId: updatedFamily
      parameters:
      - in: path
        name: id
        required: true
      - description: family request
        in: body
        name: family
        required: true
        schema:
          $ref: '#/definitions/UpdateFamilyRequest'
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/UpdateFamilyResponse'
      summary: Updates a Family.
  /growing:
    get:
      operationId: listGrowingFamilies
      parameters:
      - description: offset number of results to return
        format: int32
        in: query
        name: offset
        type: integer
      - description: maximum number of results to return
        format: int32
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/ListFastestGrowingFamiliesResponse'
      summary: Lists the fastest growing Families.
  /members:
    get:
      operationId: listMembers
      parameters:
      - description: offset number of results to return
        format: int32
        in: query
        name: offset
        type: integer
      - description: maximum number of results to return
        format: int32
        in: query
        name: limit
        type: integer
      - description: family id to filter
        format: string
        in: query
        name: family
        type: string
      - description: parent id to filter
        format: string
        in: query
        name: parent
        type: string
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/ListMembersResponse'
      summary: Lists members.
    post:
      operationId: createMembers
      parameters:
      - description: member request
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/CreateMemberRequest'
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/CreateMemberResponse'
      summary: Creates a new Member.
  /members/{id}:
    delete:
      operationId: deleteMember
      parameters:
      - description: member id
        in: path
        name: id
        required: true
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/DeleteMemberResponse'
      summary: Delete a Member.
    post:
      operationId: getMember
      parameters:
      - description: member id
        in: path
        name: id
        required: true
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/GetMemberResponse'
      summary: Gets a Member.
    put:
      operationId: updateMember
      parameters:
      - description: member id
        in: path
        name: id
        required: true
      - description: member request
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/UpdateMemberRequest'
      produces:
      - application/json
      responses:
        default:
          schema:
            $ref: '#/definitions/UpdateMemberResponse'
      summary: Update a Member.
produces:
- application/json
schemes:
- http
- https
swagger: "2.0"